enhance: extend dropdown logic to practice quizzes
